backend/api/requirements.backend.txt
-r requirements.common.txt

orjson==3.11.7
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
opentelemetry-sdk
opentelemetry-instrumentation-fastapi